Finish TAbs typing rule proof
